Update socket-close-demo: Success to produce "CLOSE-WAIT" socket
"CLOSE-WAIT" happens when the passive close side receive the active close side's FIN package but the passive close side fails to call close() and the process is still active for quite a long time.
It should be mentioned that the Socket instance in both client and the server end needs to explicitly invoke close().
If only one of the ends invokes close() then too, the socket will remain in CLOSE_WAIT state.